feat(rating): Adds rate cafea api

+++ pkg/models/ratings.go
@@ -1,8 +1,8 @@
 package models
 
 type Rating struct {
-	ID     int32   `json:"id"`
-	UserID int32   `json:"user_id"`
-	CafeID int32   `json:"cafe_id"`
-	Rating float64 `json:"rating"`
+	ID     int32 `json:"id"`
+	UserID int32 `json:"user_id"`
+	CafeID int32 `json:"cafe_id"`
+	Rating int32 `json:"rating"`
 }
